finish the ExceptionController and rebuild mybatis with mybatis-generator

+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -1,10 +1,10 @@
 CREATE TABLE question (
-  id  INT  AUTO_INCREMENT  PRIMARY KEY ,
+  id  bigint  AUTO_INCREMENT  PRIMARY KEY ,
   title  varchar(50) ,
   description  text  ,
   gmt_create  bigint  ,
   gmt_modified  bigint  ,
-  creator  INT  ,
+  creator  bigint  ,
   comment_count  INT  DEFAULT 0 ,
   view_count  INT  DEFAULT 0 ,
   like_count  INT  DEFAULT 0 ,
